+++ src/lib/compose.js
